feat: send cluster deltas with a non-sliding interval
Previously, we sent cluster deltas using a non-sliding interval.
The order was like this:
- wait for the time to send
- start sending
- finish sending
- start counting towards the next time to send.

This means that long sends will delay sending the next delta.
And if sending the delta is delayed, so is the next autoscaling cycle.

In this commit we move to sending cluster deltas using a non-sliding window.