fix: Eliminate a race condition in podmetrics informer
We were using a buffered channel to signal the "Stop" action,
but we were reading the channel at two places, so we would end up consuming the message inside the `for metrics` loop, and the `backoff.BackoffUntil` wouldn't stop.

We're fixing this by actually closing the channel instead of sending a message. Closed channels always returns `nil` on recv immediately.